Fix alias handling with default domains
