+++ BaragonData/src/main/java/com/hubspot/baragon/data/BaragonAliasDatastore.java
@@ -1,8 +1,6 @@
 package com.hubspot.baragon.data;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.utils.ZKPaths;
@@ -12,14 +10,15 @@ import com.google.common.base.Optional;
 import com.google.inject.Inject;
 import com.hubspot.baragon.config.ZooKeeperConfiguration;
 import com.hubspot.baragon.models.BaragonGroupAlias;
-import com.hubspot.baragon.models.BaragonRequest;
 
 public class BaragonAliasDatastore extends AbstractDataStore {
 
   public static final String ALIASES_ROOT = "/aliases";
 
   @Inject
-  public BaragonAliasDatastore(CuratorFramework curatorFramework, ObjectMapper objectMapper, ZooKeeperConfiguration zooKeeperConfiguration) {
+  public BaragonAliasDatastore(CuratorFramework curatorFramework,
+                               ObjectMapper objectMapper,
+                               ZooKeeperConfiguration zooKeeperConfiguration) {
     super(curatorFramework, objectMapper, zooKeeperConfiguration);
   }
 
@@ -41,36 +40,5 @@ public class BaragonAliasDatastore extends AbstractDataStore {
 
   public void deleteAlias(String name) {
     deleteNode(getAliasPath(name));
-  }
-
-  public BaragonRequest updateForAliases(BaragonRequest original) {
-    Set<String> edgeCacheDomains = new HashSet<>(original.getLoadBalancerService().getEdgeCacheDomains());
-    if (original.getLoadBalancerService().getEdgeCacheDNS().isPresent()) {
-      edgeCacheDomains.add(original.getLoadBalancerService().getEdgeCacheDNS().get());
-    }
-    return original.withUpdatedGroups(
-        processAliases(
-            original.getLoadBalancerService().getLoadBalancerGroups(),
-            original.getLoadBalancerService().getDomains(),
-            edgeCacheDomains
-        )
-    );
-  }
-
-  public BaragonGroupAlias processAliases(Set<String> groups, Set<String> domains, Set<String> edgeCacheDomains) {
-    Set<String> allGroups = new HashSet<>();
-    Set<String> allDomains = new HashSet<>(domains);
-    Set<String> allEdgeCacheDomains = new HashSet<>(edgeCacheDomains);
-    for (String group : groups) {
-      Optional<BaragonGroupAlias> maybeAlias = getAlias(group);
-      if (maybeAlias.isPresent()) {
-        allGroups.addAll(maybeAlias.get().getGroups());
-        allDomains.addAll(maybeAlias.get().getDomains());
-        allEdgeCacheDomains.addAll(maybeAlias.get().getEdgeCacheDomains());
-      } else {
-        allGroups.add(group);
-      }
-    }
-    return new BaragonGroupAlias(allGroups, allDomains, allEdgeCacheDomains);
   }
 }
